requirements.txt
pytest
pytest-testrail
PyYAML
selenium
webdriver-manager
